requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/base.in
#
anyascii==0.3.3
    # via
    #   -r requirements/base.in
    #   wagtail
appnope==0.1.4
    # via
    #   -r requirements/base.in
    #   ipython
asgiref==3.11.1
    # via
    #   -r requirements/base.in
    #   django
    #   django-allauth
asttokens==3.0.0
    # via
    #   -r requirements/base.in
    #   stack-data
attrs==25.3.0
    # via
    #   -r requirements/base.in
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via
    #   -r requirements/base.in
    #   ipython
beautifulsoup4==4.14.3
    # via
    #   -r requirements/base.in
    #   wagtail
bleach==6.2.0
    # via -r requirements/base.in
build==1.3.0
    # via -r requirements/base.in
cachetools==6.1.0
    # via -r requirements/base.in
calmjs-parse==1.3.3
    # via -r requirements/base.in
certifi==2026.1.4
    # via
    #   -r requirements/base.in
    #   requests
chardet==5.2.0
    # via -r requirements/base.in
charset-normalizer==3.4.4
    # via
    #   -r requirements/base.in
    #   requests
click==8.3.1
    # via -r requirements/base.in
csscompressor==0.9.5
    # via -r requirements/base.in
decorator==5.2.1
    # via
    #   -r requirements/base.in
    #   ipython
defusedxml==0.7.1
    # via
    #   -r requirements/base.in
    #   willow
distlib==0.4.0
    # via -r requirements/base.in
django==6.0.2
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-appconf
    #   django-components
    #   django-compressor
    #   django-filter
    #   django-guardian
    #   django-modelcluster
    #   django-permissionedforms
    #   django-stubs-ext
    #   django-taggit
    #   django-tasks
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   modelsearch
    #   wagtail
django-allauth==65.13.0
    # via -r requirements/base.in
django-appconf==1.1.0
    # via
    #   -r requirements/base.in
    #   django-compressor
django-cogwheels==0.3
    # via
    #   -r requirements/base.in
    #   wagtailmenus
django-components==0.148.0
    # via -r requirements/base.in
django-compressor==4.5.1
    # via -r requirements/base.in
django-environ==0.12.0
    # via -r requirements/base.in
django-filter==25.2
    # via
    #   -r requirements/base.in
    #   wagtail
django-guardian==3.0.3
    # via -r requirements/base.in
django-modelcluster==6.4.1
    # via
    #   -r requirements/base.in
    #   wagtail
django-permissionedforms==0.1
    # via
    #   -r requirements/base.in
    #   wagtail
django-stubs-ext==5.2.9
    # via
    #   -r requirements/base.in
    #   django-tasks
django-taggit==6.1.0
    # via
    #   -r requirements/base.in
    #   wagtail
django-tasks==0.9.0
    # via
    #   -r requirements/base.in
    #   modelsearch
    #   wagtail
django-treebeard==4.8.0
    # via
    #   -r requirements/base.in
    #   wagtail
djangorestframework==3.16.1
    # via
    #   -r requirements/base.in
    #   wagtail
djc-core==1.3.0
    # via
    #   -r requirements/base.in
    #   django-components
draftjs-exporter==5.2.0
    # via
    #   -r requirements/base.in
    #   wagtail
emoji==2.15.0
    # via -r requirements/base.in
et-xmlfile==2.0.0
    # via
    #   -r requirements/base.in
    #   openpyxl
executing==2.2.0
    # via
    #   -r requirements/base.in
    #   stack-data
fastjsonschema==2.21.1
    # via -r requirements/base.in
filelock==3.20.3
    # via -r requirements/base.in
filetype==1.2.0
    # via
    #   -r requirements/base.in
    #   willow
heroicons==2.11.0
    # via -r requirements/base.in
identify==2.6.12
    # via -r requirements/base.in
idna==3.11
    # via
    #   -r requirements/base.in
    #   requests
importlib-metadata==8.7.0
    # via -r requirements/base.in
ipython==8.12.3
    # via -r requirements/base.in
jedi==0.19.2
    # via
    #   -r requirements/base.in
    #   ipython
jinja2==3.1.6
    # via -r requirements/base.in
jsonschema==4.25.0
    # via -r requirements/base.in
jsonschema-specifications==2025.4.1
    # via
    #   -r requirements/base.in
    #   jsonschema
laces==0.1.2
    # via
    #   -r requirements/base.in
    #   wagtail
markdown==3.8.2
    # via -r requirements/base.in
markupsafe==3.0.2
    # via
    #   -r requirements/base.in
    #   jinja2
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via -r requirements/base.in
mistune==3.1.3
    # via -r requirements/base.in
modelsearch==1.1.1
    # via
    #   -r requirements/base.in
    #   wagtail
nodeenv==1.9.1
    # via -r requirements/base.in
openpyxl==3.1.5
    # via
    #   -r requirements/base.in
    #   wagtail
packaging==25.0
    # via
    #   build
    #   wheel
parso==0.8.4
    # via
    #   -r requirements/base.in
    #   jedi
pexpect==4.9.0
    # via
    #   -r requirements/base.in
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements/base.in
    #   ipython
pillow==12.1.0
    # via
    #   -r requirements/base.in
    #   pillow-heif
    #   wagtail
pillow-heif==1.2.0
    # via
    #   -r requirements/base.in
    #   willow
pluggy==1.6.0
    # via -r requirements/base.in
ply==3.11
    # via
    #   -r requirements/base.in
    #   calmjs-parse
prompt-toolkit==3.0.51
    # via
    #   -r requirements/base.in
    #   ipython
ptyprocess==0.7.0
    # via
    #   -r requirements/base.in
    #   pexpect
pure-eval==0.2.3
    # via
    #   -r requirements/base.in
    #   stack-data
pygments==2.19.2
    # via ipython
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via -r requirements/base.in
pytz==2025.2
    # via -r requirements/base.in
pyzmq==27.0.0
    # via -r requirements/base.in
rcssmin==1.1.2
    # via
    #   -r requirements/base.in
    #   django-compressor
referencing==0.36.2
    # via
    #   -r requirements/base.in
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via -r requirements/base.in
requests==2.32.5
    # via
    #   -r requirements/base.in
    #   stripe
    #   wagtail
    #   yarg
rjsmin==1.2.2
    # via
    #   -r requirements/base.in
    #   django-compressor
rpds-py==0.26.0
    # via
    #   -r requirements/base.in
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via
    #   -r requirements/base.in
    #   beautifulsoup4
sqlparse==0.5.5
    # via
    #   -r requirements/base.in
    #   django
stack-data==0.6.3
    # via
    #   -r requirements/base.in
    #   ipython
stripe==12.3.0
    # via -r requirements/base.in
telepath==0.3.1
    # via
    #   -r requirements/base.in
    #   wagtail
tinycss2==1.4.0
    # via -r requirements/base.in
tornado==6.5.1
    # via -r requirements/base.in
traitlets==5.14.3
    # via
    #   -r requirements/base.in
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   -r requirements/base.in
    #   beautifulsoup4
    #   django-components
    #   django-stubs-ext
    #   django-tasks
    #   referencing
    #   stripe
uritemplate==4.2.0
    # via -r requirements/base.in
urllib3==2.6.3
    # via
    #   -r requirements/base.in
    #   requests
wagtail==7.3
    # via
    #   -r requirements/base.in
    #   wagtailmenus
wagtailmenus @ git+https://github.com/baldwinboy/wagtailmenus@master
    # via -r requirements/base.in
wcwidth==0.2.13
    # via
    #   -r requirements/base.in
    #   prompt-toolkit
webencodings==0.5.1
    # via
    #   -r requirements/base.in
    #   bleach
    #   tinycss2
wheel==0.46.2
    # via -r requirements/base.in
willow[heif]==1.12.0
    # via
    #   -r requirements/base.in
    #   wagtail
yarg==0.1.9
    # via -r requirements/base.in
zipp==3.23.0
    # via
    #   -r requirements/base.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
